revert question util imports
