+++ src/components/Squeak/components/QuestionForm.tsx
@@ -10,7 +10,7 @@ import { usePost } from 'components/PostLayout/hooks'
 import qs from 'qs'
 import OSButton from 'components/OSButton'
 import uploadImage from '../util/uploadImage'
-import { fetchTopicGroups, topicGroupsSorted } from '../index'
+import { fetchTopicGroups, topicGroupsSorted } from '../../../pages/questions'
 import usePostHog from 'hooks/usePostHog'
 import { navigate } from 'gatsby'
 import { useAppStatus } from 'hooks/useAppStatus'
